Started Composer project setup.
PSR-4 namespaces, PSR-2 code style.

Included jQuery and path-data-polyfill distributives in scripts/vendor.
