+++ src/Mvc/Model.php
@@ -1,0 +1,24 @@
+<?php
+namespace HeraldryEngine\Mvc;
+
+/**
+ * A model.
+ */
+class Model
+{
+	/**
+	 * Create a new model.
+	 */
+	public function __construct()
+	{
+
+	}
+
+	/**
+	 * Prepare the model.
+	 */
+	public function prepareModel()
+	{
+
+	}
+}
