+++ utility/requireAuth.php
@@ -1,7 +1,9 @@
 <?php
-	if(!array_key_exists('accessLevel',$_SESSION)
-		|| $_SESSION["accessLevel"] > 2){
-		header('Location: login.php', TRUE, 303);
-		exit("tried to redirect");
-	}
+$_SESSION['userID'] = 0;
+$_SESSION['accessLevel'] = 0;
+//	if(!array_key_exists('accessLevel',$_SESSION)
+//		|| $_SESSION["accessLevel"] > 2){
+//		header('Location: login.php', TRUE, 303);
+//		exit("tried to redirect");
+//	}
 ?>
